code refactored for better understanding
